# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/admin/sources/{id}/kill-switch": {
      "post": {
        "operationId": "AdminController_kill",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 500
                  }
                },
                "required": ["reason"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Immediately disable a source (hides its prices from public views).",
        "tags": ["admin"]
      }
    },
    "/v1/admin/sources/{id}/kill-switch/release": {
      "post": {
        "operationId": "AdminController_release",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 500
                  }
                },
                "required": ["reason"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["admin"]
      }
    },
    "/v1/admin/sources/{id}/changes": {
      "post": {
        "operationId": "AdminController_requestChange",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 500
                  },
                  "change": {
                    "type": "object",
                    "properties": {
                      "legal_status": {
                        "type": "string",
                        "enum": ["green", "amber", "red", "disabled"]
                      },
                      "approval_ref": {
                        "type": "string",
                        "maxLength": 300
                      },
                      "tos_archive_url": {
                        "type": "string",
                        "maxLength": 500
                      },
                      "robots_archive_url": {
                        "type": "string",
                        "maxLength": 500
                      },
                      "refresh_cron": {
                        "type": "string",
                        "maxLength": 100
                      }
                    },
                    "additionalProperties": false
                  }
                },
                "required": ["reason", "change"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["admin"]
      }
    },
...
    "/v1/admin/matches/{id}/decide": {
      "post": {
        "operationId": "AdminController_decide",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "product_id": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "required": ["product_id"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["admin"]
      }
    },
...
    "/v1/admin/products/merge": {
      "post": {
        "operationId": "AdminController_merge",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "keep": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "drop": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "required": ["keep", "drop"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["admin"]
      }
    },
...
    "/v1/admin/takedowns": {
      "get": {
        "operationId": "AdminController_takedowns",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["admin"]
      },
      "post": {
        "operationId": "AdminController_createTakedown",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "retailer_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "source_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "requester_name": {
                    "type": "string",
                    "maxLength": 200
                  },
                  "requester_email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 254
                  },
                  "channel": {
                    "type": "string",
                    "enum": ["email", "web_form", "letter", "other"],
                    "default": "email"
                  },
                  "summary": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 2000
                  }
                },
                "required": ["summary"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["admin"]
      }
    },
...
    "/v1/admin/takedowns/{id}/resolve": {
      "post": {
        "operationId": "AdminController_resolve",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "source_disabled",
                      "price_corrected",
                      "content_hidden",
                      "rejected",
                      "forwarded_to_counsel"
                    ]
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 2000
                  }
                },
                "required": ["action"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["admin"]
      }
    },
...
  },
  "info": {
    "title": "Qarib API",
    "description": "Qatar grocery price comparison. Prices come from retailers, partners and community reports and may change.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "qarib_at"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}